PCHR-1421: PR Updates
