+++ uk.co.compucorp.civicrm.hrcore/backstop_data/casper_scripts/ssp-leave-absences/staff/my-leave-calendar_current-month-visible.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var page = require('../../../page-objects/ssp-la-my-leave');
+var page = require('../../../page-objects/ssp-leave-absences-my-leave-calendar');
 
 module.exports = function (casper) {
   page.init(casper);
