+++ uk.co.compucorp.civicrm.hrcore/backstop_data/casper_scripts/ssp-leave-absences/staff/my-report-pending-show-comments.js
@@ -1,0 +1,14 @@
+'use strict';
+
+var page = require('../../../page-objects/ssp-leave-absences-my-leave-report');
+
+// precondition: need to have the first leave request on the pending list with at least a comment
+module.exports = function (casper) {
+  var myReport = page.init(casper);
+  myReport.openSection('\'pending\'');
+  myReport.openActions();
+
+  myReport.editRequest().then(function (modal) {
+    modal.selectTab('\'Comments\'');
+  });
+};
