+++ uk.co.compucorp.civicrm.hrleaveandabsences/js/angular/src/leave-absences/shared/controllers/request-ctrl.js
@@ -28,7 +28,7 @@ define([
         initialLeaveRequestAttributes = {}, //used to compare the change in leaverequest in edit mode
         mode = '', //can be edit, create, view
         role = '', //could be manager, owner or admin
-        originalCommentsLength = 0; //number of comments when the request model is loaded
+        initialCommentsLength = 0; //number of comments when the request model is loaded
 
       this.absencePeriods = [];
       this.absenceTypes = [];
@@ -194,7 +194,7 @@ define([
         if (this.isMode('edit')) {
           canSubmit = canSubmit && !_.isEqual(initialLeaveRequestAttributes, this.request.attributes());
           //user has added a comment
-          canSubmit = canSubmit || this.request.comments.length != originalCommentsLength;
+          canSubmit = canSubmit || this.request.comments.length !== initialCommentsLength;
         }
 
         //check if manager has changed status
@@ -644,7 +644,7 @@ define([
                 self.uiOptions.multipleDays = false;
               }
 
-              originalCommentsLength = self.request.comments.length;
+              initialCommentsLength = self.request.comments.length;
             }
           });
       };
