+++ uk.co.compucorp.civicrm.hrleaveandabsences/views/manager-leave/components/manager-leave-requests.html
@@ -113,16 +113,18 @@
               <td style="width: 25%;">{{request.from_date}} - {{request.to_date}}</td>
               <td style="width: 20%;">{{ctrl.getLeaveStatusByValue(request.status_id)}}</td>
               <td style="width: 10%;" class="text-center">
-                <div class="btn-group">
-                  <button type="button"
-                          class="btn btn-default dropdown-toggle btn-inline"
-                          data-toggle="dropdown"
-                          aria-haspopup="true"
-                          aria-expanded="false">
-                    <i class="fa fa-ellipsis-v" aria-hidden="true"></i>
-                  </button>
-                  <ul class="dropdown-menu">
-                    <li><a href='#' leave-request-popup leave-request='request' contact-id="ctrl.contactId">Respond</a></li>
+                <div uib-dropdown dropdown-append-to-body>
+                  <a href class="btn btn-default btn-inline" uib-dropdown-toggle>
+                    <i class="fa fa-ellipsis-v"></i>
+                  </a>
+                  <ul class="dropdown-menu dropdown-menu-right" uib-dropdown-menu>
+                    <li>
+                      <a
+                        href='#'
+                        leave-request-popup
+                        leave-request='request'
+                        contact-id="ctrl.contactId">Respond</a>
+                    </li>
                   </ul>
                 </div>
               </td>
