+++ uk.co.compucorp.civicrm.hrleaveandabsences/views/shared/directives/leave-request-popup.html
@@ -193,7 +193,7 @@
                 class="chr_wysiwyg"
                 text-angular
                 prevent-animations
-                ng-disabled="$ctrl.isMode('view')"
+                ta-disabled="$ctrl.isMode('view')"
                 ta-toolbar="[['bold','italics','underline']]"
                 ng-model="$ctrl.comment.text"></div>
               <div class="chr_wysiwyg__action">
